Implement redux-thunk
