+++ package.json
@@ -17,6 +17,7 @@
     "@types/axios": "^0.14.0",
     "@types/react-redux": "^7.1.9",
     "@types/redux": "^3.6.0",
+    "@types/redux-thunk": "^2.1.0",
     "axios": "^0.21.0",
     "expo": "~38.0.8",
     "expo-font": "~8.2.1",
@@ -31,7 +32,8 @@
     "react-native-screens": "~2.9.0",
     "react-native-web": "~0.11.7",
     "react-redux": "^7.2.1",
-    "redux": "^4.0.5"
+    "redux": "^4.0.5",
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
