+++ src/containers/NotificationDetail.tsx
@@ -1,5 +1,5 @@
 import { connect, MapStateToProps } from 'react-redux';
-import { AppState } from '../store/reducers';
+import { AppState } from '../store';
 import NotificationDetail, {
   NotificationDetailProps,
 } from '../screens/NotificationDetail/NotificationDetail';
